using useEndpoint in Maintenance Screen
