+++ src/client/containers/MaintenanceContainer.jsx
@@ -1,27 +1,29 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { useConnectionString } from 'context/withConnectionString'
 import MaintenanceScreen from 'pages/Maintenance/MaintenanceScreen.jsx'
-import useEnvironment from 'hooks/useEnvironment'
+import useEndpoint from 'hooks/useEndpoint'
 import Main from 'Main'
 
+const devEnv = {
+  env: 'development',
+  maintainance: false,
+  protocol: 'http',
+  host: '<firebase functions url goes here>',
+}
 
 function MaintenanceContainer({ history }) {
-  const connectionString = useConnectionString()
-  const [env, err] = useEnvironment(connectionString)
+  const [err, loading, env] = useEndpoint({
+    path: '/environment',
+  }, devEnv)
 
-  if (err) {
-    return <MaintenanceScreen />
-  }
+  if (loading) return null
 
-  if (process.env.NODE_ENV !== 'production') {
-    // Changes this is you want to see the MaintenanceScreen
-    return <Main history={history} />
-  }
+  if (err) return <MaintenanceScreen />
 
-  if (env.maintainance !== 'true') {
-    return <Main history={history} />
-  }
+  // Changes this is you want to see the MaintenanceScreen
+  if (process.env.NODE_ENV !== 'production') return <Main history={history} />
+
+  if (env.maintainance !== 'true') return <Main history={history} />
 
   return <MaintenanceScreen />
 }
